apollo.sh: add older script to avoid weird script errors
* temporarily

+++ apollo.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Apollo Build Script V3.5
+# DS Build Script V3.0
 # For Exynos9810
 # Forked from Exynos8890 Script
 # Coded by AnanJaser1211 @ 2019-2022
@@ -17,12 +17,8 @@
 # limitations under the License.
 # Main Dir
 CR_DIR=$(pwd)
-# Compiler Dir
-CR_TC=../compiler
-# Target ARCH
-CR_ARCH=arm64
 # Define proper arch and dir for dts files
-CR_DTS=arch/$CR_ARCH/boot/dts/exynos
+CR_DTS=arch/arm64/boot/dts/exynos
 # Define boot.img out dir
 CR_OUT=$CR_DIR/DS/Out
 CR_PRODUCT=$CR_DIR/DS/Product
@@ -34,11 +30,9 @@ CR_AIK=$CR_DIR/DS/A.I.K
 # Main Ramdisk Location
 CR_RAMDISK=$CR_DIR/DS/Ramdisk
 # Compiled image name and location (Image/zImage)
-CR_KERNEL=$CR_DIR/arch/$CR_ARCH/boot/Image
+CR_KERNEL=$CR_DIR/arch/arm64/boot/Image
 # Compiled dtb by dtbtool
-CR_DTB=$CR_DIR/arch/$CR_ARCH/boot/dtb.img
-# defconfig dir
-CR_DEFCONFIG=$CR_DIR/arch/$CR_ARCH/configs
+CR_DTB=$CR_DIR/arch/arm64/boot/dtb.img
 # Kernel Name and Version
 CR_VERSION=V2.0
 CR_NAME=DS-ACK
@@ -47,8 +41,10 @@ CR_JOBS=$(nproc --all)
 # Target Android version
 CR_ANDROID=q
 CR_PLATFORM=13.0.0
+# Target ARCH
+CR_ARCH=arm64
 # Current Date
-CR_DATE=$(date +%d.%m.%Y)
+CR_DATE=$(date +%Y%m%d)
 # General init
 export ANDROID_MAJOR_VERSION=$CR_ANDROID
 export PLATFORM_VERSION=$CR_PLATFORM
@@ -72,125 +68,66 @@ CR_CONFIG_SPLIT=NULL
 CR_CONFIG_DS=ds_defconfig
 CR_CONFIG_INTL=eur_defconfig
 CR_CONFIG_KOR=kor_defconfig
-CR_SELINUX="2"
+CR_SELINUX="1"
 CR_KSU="n"
 CR_CLEAN="n"
-# Default Compilation
-DEFAULT_TARGET=3   # crownlte
-DEFAULT_COMPILER=3 # clang18
-DEFAULT_SELINUX=2  # enforce
-DEFAULT_KSU=y      # enabled
-DEFAULT_CLEAN=n    # dirty
+# Compiler Paths
+CR_GCC4=~/Android/Toolchains/aarch64-linux-android-4.9/bin/aarch64-linux-android-
+CR_GCC9=~/compiler/aarch64-linux-gnu-9.x/bin/aarch64-linux-gnu-
+CR_GCC12=~/Android/Toolchains/aarch64-linux-gnu-12.x/bin/aarch64-linux-gnu-
+CR_GCC13=~/Android/Toolchains/aarch64-linux-gnu-13.x/bin/aarch64-linux-gnu-
+CR_CLANG_11=~/Android/Toolchains/clang-r383902-jopp
+CR_CLANG_12=~/compiler/clang-12.0.4-r416183
+CR_CLANG_14=~/Android/Toolchains/clang-r450784d
+CR_CLANG_EXP=~/Android/Toolchains/clang-r522817
 #####################################################
 
 # Compiler Selection
 BUILD_COMPILER()
 {
-
-# Auto download and setup compilers
-# For manually adding compiler, add it under
-# Apollo/toolchain/clang-custom and select option 7
-
-# Clang Versions and features
-
 if [ $CR_COMPILER = "1" ]; then
-CR_CLANG_URL=https://android.googlesource.com/platform/prebuilts/clang/host/linux-x86/+archive/70d66e12a35fefd55ca4cf821f35405de6ffbe14/clang-r349610.tar.gz
-CR_CLANG=$CR_TC/clang-8.0.8-r349610
+export CROSS_COMPILE=$CR_GCC4
+compile="make"
+CR_COMPILER="$CR_GCC4"
 fi
 if [ $CR_COMPILER = "2" ]; then
-CR_CLANG_URL=https://android.googlesource.com/platform/prebuilts/clang/host/linux-x86/+archive/refs/heads/llvm-r416183/clang-r416183.tar.gz
-CR_CLANG=$CR_TC/clang-12.0.4-r416183
+export CROSS_COMPILE=$CR_GCC9
+compile="make"
+CR_COMPILER="$CR_GCC9"
 fi
 if [ $CR_COMPILER = "3" ]; then
-CR_CLANG_URL=https://android.googlesource.com/platform/prebuilts/clang/host/linux-x86/+archive/refs/heads/llvm-r450784/clang-r450784b.tar.gz
-CR_CLANG=$CR_TC/clang-14.0.4-r450784
+export CROSS_COMPILE=$CR_GCC12
+compile="make"
+CR_COMPILER="$CR_GCC12"
 fi
 if [ $CR_COMPILER = "4" ]; then
-CR_CLANG_URL=https://android.googlesource.com/platform/prebuilts/clang/host/linux-x86/+archive/refs/heads/llvm-r522817/clang-r522817.tar.gz
-CR_CLANG=$CR_TC/clang-18.0.1-r522817
+export CROSS_COMPILE=$CR_GCC13
+compile="make"
+CR_COMPILER="$CR_GCC13"
 fi
-if [ $CR_COMPILER = "5" ]; then
-CR_CLANG_URL=https://android.googlesource.com/platform/prebuilts/clang/host/linux-x86/+archive/refs/heads/llvm-r530567/clang-r530567.tar.gz
-CR_CLANG=$CR_TC/clang-19.0.0-r530567
-fi
-if [ $CR_COMPILER = "6" ]; then
-CR_CLANG_URL=https://github.com/Neutron-Toolchains/clang-build-catalogue/releases/download/05012024/neutron-clang-05012024.tar.zst
-CR_CLANG=$CR_TC/neutron-clang-18.0.0
-fi
-if [ $CR_COMPILER = "7" ]; then
-CR_CLANG_URL=https://github.com/Neutron-Toolchains/clang-build-catalogue/releases/download/10032024/neutron-clang-10032024.tar.zst
-CR_CLANG=$CR_TC/neutron-clang-19.0.0
-fi
-if [ $CR_COMPILER = "8" ]; then
-CR_CLANG=$CR_TC/neutron-clang-20.0.0
-fi
-if [ $CR_COMPILER = "9" ]; then
-CR_CLANG=$CR_TC/clang-custom
-fi
-
-if [ $CR_COMPILER != "9" ]; then
-	if [ ! -d "$CR_CLANG/bin" ]; then
-		echo " "
-		echo " $CR_CLANG compiler is missing"
-		echo " "
-		echo " "
-		read -p "Download Toolchain ? (y/n) > " TC_DL
-		if [ $TC_DL = "y" ]; then
-			echo "Checking URL validity..."
-			URL=$CR_CLANG_URL
-			if curl --output /dev/null --silent --head --fail "$URL"; then
-				echo "URL exists: $URL"
-				echo "Downloading $CR_CLANG"
-				if [ ! -e $CR_TC ]; then
-					mkdir $CR_TC
-				fi
-				if [ ! -e $CR_CLANG ]; then
-					mkdir $CR_CLANG
-				else
-					# Remove incomplete
-					rm -rf $CR_CLANG
-					mkdir $CR_CLANG
-				fi
-				wget -qO- $URL | tar --use-compress-program=unzstd -xv -C $CR_CLANG
-				if [ $? -ne 0 ]; then
-					echo "Download failed or was incomplete"
-					echo "Setup Compiler and try again"
-					exit 0;
-				fi
-				# Neutron Needs patches
-				if [ $CR_COMPILER = "5" ] || [ $CR_COMPILER = "6" ]; then
-					cd $CR_CLANG
-					bash <(curl -s "https://raw.githubusercontent.com/Neutron-Toolchains/antman/main/antman") --patch=glibc
-					cd $CR_DIR
-				fi
-				echo "Compiler Downloaded."
-			else
-				echo "Invalid URL: $URL"
-				exit 0;
-			fi
-		else
-			echo " Aborting "
-			echo " Setup Compiler and try again"
-			exit 0;
-		fi
-	fi
-else
-    if [ ! -d "$CR_CLANG/bin" ] || [ ! -d "$CR_CLANG/lib" ]; then
-        echo "clang-custom compiler is missing in $CR_TC/clang-custom"
-        exit 0;
+if [ $CR_COMPILER = "5" ] || [ $CR_COMPILER = "6" ] || [ $CR_COMPILER = "7" ] || [ $CR_COMPILER = "8" ]; then
+    if [ $CR_COMPILER = "5" ]; then
+        CR_CLANG=$CR_CLANG_11
+    elif [ $CR_COMPILER = "6" ]; then
+        CR_CLANG=$CR_CLANG_12
+    elif [ $CR_COMPILER = "7" ]; then
+        CR_CLANG=$CR_CLANG_14
+    elif [ $CR_COMPILER = "8" ]; then
+        CR_CLANG=$CR_CLANG_EXP
     fi
-fi
 
-# Clang Features (18 and higher)
-if [ $CR_COMPILER -ge 3 ]; then
-export CONFIG_THINLTO=y
-export CONFIG_UNIFIEDLTO=y
-export CONFIG_LLVM_MLGO_REGISTER=y
-export CONFIG_LLVM_POLLY=y
-export CONFIG_LLVM_DFA_JUMP_THREAD=y
+# Check packages
+for pkg in gcc-arm-linux-gnueabi gcc-aarch64-linux-gnu; do
+if ! dpkg-query -W -f='${Status}' $pkg | grep "ok installed"; then
+    echo " $pkg is missing, please install with sudo apt-get install $pkg"
+    exit 0;
 fi
+done
 
 export PATH=$CR_CLANG/bin:$CR_CLANG/lib:${PATH}
+export CLANG_TRIPLE=$CR_GCC9
+export CROSS_COMPILE=$CR_GCC9
+export CROSS_COMPILE_ARM32=arm-linux-gnueabi-
 export CC=$CR_CLANG/bin/clang
 export REAL_CC=$CR_CLANG/bin/clang
 export LD=$CR_CLANG/bin/ld.lld
@@ -201,17 +138,20 @@ export OBJDUMP=$CR_CLANG/bin/llvm-objdump
 export READELF=$CR_CLANG/bin/llvm-readelf
 export STRIP=$CR_CLANG/bin/llvm-strip
 export LLVM=1
-export KALLSYMS_EXTRA_PASS=1
+export LLVM_IAS=1
 export ARCH=arm64 && export SUBARCH=arm64
 compile="make ARCH=arm64 CC=clang"
-CR_COMPILER_ARG="$CR_CLANG"
+CR_COMPILER="$CR_CLANG"
+fi
 }
 
 # Clean-up Function
 
 BUILD_CLEAN()
 {
-if [[ "$CR_CLEAN" =~ ^[yY]$ ]]; then
+if [ $CR_CLEAN = "y" ]; then
+     echo " "
+     echo " Cleaning build dir"
      $compile clean && $compile mrproper
      rm -r -f $CR_DTB
      rm -r -f $CR_KERNEL
@@ -221,7 +161,10 @@ if [[ "$CR_CLEAN" =~ ^[yY]$ ]]; then
      rm -rf $CR_DIR/.config
      rm -rf $CR_OUT/*.img
      rm -rf $CR_OUT/*.zip
-else
+fi
+if [ $CR_CLEAN = "n" ]; then
+     echo " "
+     echo " Skip Full cleaning"
      rm -r -f $CR_DTB
      rm -r -f $CR_KERNEL
      rm -rf $CR_DTS/.*.tmp
@@ -241,110 +184,58 @@ BUILD_IMAGE_NAME()
     
 }
 
-# Build options
-BUILD_OPTIONS()
-{
-	# KSU Version
-	KSU_VERSION=$( [ -f "drivers/kernelsu/Makefile" ] && grep -oP '(?<=-DKSU_VERSION=)[0-9]+' drivers/kernelsu/Makefile )
-	echo "----------------------------------------------"
-	echo " DS Kernel Build Options "
-	echo " "
-	echo " Kernel		- $CR_IMAGE_NAME"
-	echo " Device		- $CR_VARIANT"
-	echo " Compiler	- $CR_COMPILER_ARG"
-	if [[ "$CR_CLEAN" =~ ^[yY]$ ]]; then
-		echo " Env		- Clean Build"
-	else
-		echo " Env		- Dirty Build"
-	fi
-	if [ $CR_SELINUX = "1" ]; then
-		echo " SELinux	- Permissive"
-	else
-		echo " SELinux	- Enforcing"
-	fi
-	if [[ "$CR_KSU" =~ ^[yY]$ ]]; then
-		if [ -n "$KSU_VERSION" ]; then
-		echo " KernelSU	- Version: $KSU_VERSION"
-		else
-		echo " KernelSU	- Enabled"
-		fi
-	else
-		echo " KernelSU	- Disabled"
-	fi
-	echo " "
-}
-
 # Config Generation Function
 
 BUILD_GENERATE_CONFIG()
 {
   # Only use for devices that are unified with 2 or more configs
   echo "----------------------------------------------"
-  echo " Generating defconfig for $CR_VARIANT"
+  echo "Building defconfig for $CR_VARIANT"
   echo " "
   # Respect CLEAN build rules
   BUILD_CLEAN
-  if [ -e $CR_DEFCONFIG/tmp_defconfig ]; then
-    echo " Clean-up old config "
-    rm -rf $CR_DEFCONFIG/tmp_defconfig
+  if [ -e $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig ]; then
+    echo " cleanup old configs "
+    rm -rf $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   fi
-  echo " Base	- $CR_CONFIG "
-  cp -f $CR_DEFCONFIG/$CR_CONFIG $CR_DEFCONFIG/tmp_defconfig
+  echo " Copy $CR_CONFIG "
+  cp -f $CR_DIR/arch/$CR_ARCH/configs/$CR_CONFIG $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   # Split-config support for devices with unified defconfigs (Universal + device)
   if [ $CR_CONFIG_SPLIT = NULL ]; then
     echo " No split config support! "
   else
-    echo " Device - $CR_CONFIG_SPLIT "
-    cat $CR_DEFCONFIG/$CR_CONFIG_SPLIT >> $CR_DEFCONFIG/tmp_defconfig
+    echo " Copy $CR_CONFIG_SPLIT "
+    cat $CR_DIR/arch/$CR_ARCH/configs/$CR_CONFIG_SPLIT >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   fi
   # Regional Config
-  echo " Region	- $CR_CONFIG_REGION "
-  cat $CR_DEFCONFIG/$CR_CONFIG_REGION >> $CR_DEFCONFIG/tmp_defconfig
+  echo " Copy $CR_CONFIG_REGION "
+  cat $CR_DIR/arch/$CR_ARCH/configs/$CR_CONFIG_REGION >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   # DS Custom defconfig
-  echo " DS	- $CR_CONFIG_DS "
-  cat $CR_DEFCONFIG/$CR_CONFIG_DS >> $CR_DEFCONFIG/tmp_defconfig
+  echo " Copy $CR_CONFIG_DS "
+  cat $CR_DIR/arch/$CR_ARCH/configs/$CR_CONFIG_DS >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   # Selinux Never Enforce all targets
   if [ $CR_SELINUX = "1" ]; then
-    echo " Building SELinux Permissive Kernel"
-    echo "CONFIG_ALWAYS_PERMISSIVE=y" >> $CR_DEFCONFIG/tmp_defconfig
+    echo " Building Permissive Kernel"
+    echo "CONFIG_ALWAYS_PERMISSIVE=y" >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
     CR_IMAGE_NAME=$CR_IMAGE_NAME-Permissive
     zver=$zver-Permissive
-  else
-    echo " Building SELinux Enforced Kernel"
   fi
-  if [[ "$CR_KSU" =~ ^[yY]$ ]]; then
-    echo " Building KernelSU"
-    echo "CONFIG_KSU=y" >> $CR_DEFCONFIG/tmp_defconfig
-    CR_IMAGE_NAME=$CR_IMAGE_NAME-KSU
+  if [ $CR_KSU = "y" ]; then
+    echo " Building KernelSU Kernel"
+    echo "CONFIG_KSU=y" >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
+    CR_IMAGE_NAME=$CR_IMAGE_NAME-ksu
     zver=$zver-KernelSU
   else
-    echo "# CONFIG_KSU is not set" >> $CR_DEFCONFIG/tmp_defconfig
+    echo "# CONFIG_KSU is not set" >> $CR_DIR/arch/$CR_ARCH/configs/tmp_defconfig
   fi
-  echo " $CR_VARIANT config generated "
-  echo " "
+  echo " Set $CR_VARIANT to generated config "
   CR_CONFIG=tmp_defconfig
 }
 
 # Kernel information Function
 BUILD_OUT()
 {
-# KSU Version
-	KSU_VERSION=$( [ -f "drivers/kernelsu/Makefile" ] && grep -oP '(?<=-DKSU_VERSION=)[0-9]+' drivers/kernelsu/Makefile )
-  echo "----------------------------------------------"
-  echo " Kernel		- $CR_IMAGE_NAME"
-  echo " Device		- $CR_VARIANT"
-  echo " Compiler	- $CR_COMPILER_ARG"
-	if [[ "$CR_CLEAN" =~ ^[yY]$ ]]; then
-		echo " Env		- Clean Build"
-	else
-		echo " Env		- Dirty Build"
-	fi
-	if [ $CR_SELINUX = "1" ]; then
-		echo " SELinux	- Permissive"
-	else
-		echo " SELinux	- Enforcing"
-	fi
-  echo " KernelSU	- Version: $KSU_VERSION"
+  echo " "
   echo "----------------------------------------------"
   echo "$CR_VARIANT kernel build finished."
   echo "Compiled DTB Size = $sizdT Kb"
@@ -364,7 +255,7 @@ BUILD_ZIMAGE()
 	export LOCALVERSION=-$CR_IMAGE_NAME
 	echo "Make $CR_CONFIG"
 	$compile $CR_CONFIG
-	echo "Make Kernel with $CR_COMPILER_ARG"
+	echo "Make Kernel with $CR_COMPILER"
 	$compile -j$CR_JOBS
 	if [ ! -e $CR_KERNEL ]; then
 	exit 0;
@@ -447,44 +338,44 @@ BUILD()
 	if [ "$CR_TARGET" = "1" ]; then
 		echo " Galaxy S9 INTL"
 		CR_CONFIG_SPLIT=$CR_CONFIG_G960
-		CR_CONFIG_REGION=$CR_CONFIG_INTL
+        CR_CONFIG_REGION=$CR_CONFIG_INTL
 		CR_VARIANT=$CR_VARIANT_G960F
 		export "CONFIG_MACH_EXYNOS9810_STARLTE_EUR_OPEN=y"
 	fi
 	if [ "$CR_TARGET" = "2" ]; then
 		echo " Galaxy S9+ INTL"
 		CR_CONFIG_SPLIT=$CR_CONFIG_G965
-		CR_CONFIG_REGION=$CR_CONFIG_INTL
+        CR_CONFIG_REGION=$CR_CONFIG_INTL
 		CR_VARIANT=$CR_VARIANT_G965F
 		export "CONFIG_MACH_EXYNOS9810_STAR2LTE_EUR_OPEN=y"
 	fi
 	if [ "$CR_TARGET" = "3" ]
 	then
-		echo " Galaxy Note9 INTL"
+		echo " Galaxy Note 9 INTL"
 		CR_CONFIG_SPLIT=$CR_CONFIG_N960
-		CR_CONFIG_REGION=$CR_CONFIG_INTL
+        CR_CONFIG_REGION=$CR_CONFIG_INTL
 		CR_VARIANT=$CR_VARIANT_N960F
 		export "CONFIG_MACH_EXYNOS9810_CROWNLTE_EUR_OPEN=y"
 	fi
 	if [ "$CR_TARGET" = "4" ]; then
 		echo " Galaxy S9 KOR"
 		CR_CONFIG_SPLIT=$CR_CONFIG_G960
-		CR_CONFIG_REGION=$CR_CONFIG_KOR
+        CR_CONFIG_REGION=$CR_CONFIG_KOR
 		CR_VARIANT=$CR_VARIANT_G960N
 		export "CONFIG_MACH_EXYNOS9810_STARLTE_KOR=y"
 	fi
 	if [ "$CR_TARGET" = "5" ]; then
 		echo " Galaxy S9+ KOR"
 		CR_CONFIG_SPLIT=$CR_CONFIG_G965
-		CR_CONFIG_REGION=$CR_CONFIG_KOR
+        CR_CONFIG_REGION=$CR_CONFIG_KOR
 		CR_VARIANT=$CR_VARIANT_G965N
 		export "CONFIG_MACH_EXYNOS9810_STAR2LTE_KOR=y"
 	fi
 	if [ "$CR_TARGET" = "6" ]
 	then
-		echo " Galaxy Note9 KOR"
+		echo " Galaxy Note 9 KOR"
 		CR_CONFIG_SPLIT=$CR_CONFIG_N960
-		CR_CONFIG_REGION=$CR_CONFIG_KOR
+        CR_CONFIG_REGION=$CR_CONFIG_KOR
 		CR_VARIANT=$CR_VARIANT_N960N
 		export "CONFIG_MACH_EXYNOS9810_CROWNLTE_KOR=y"
 	fi	
@@ -493,8 +384,6 @@ BUILD()
 	BUILD_CLEAN
 	BUILD_IMAGE_NAME
 	BUILD_GENERATE_CONFIG
-	# Print build options
-	BUILD_OPTIONS
 	BUILD_ZIMAGE
 	BUILD_DTB
 	if [ "$CR_MKZIP" = "y" ]; then # Allow Zip Package for mass compile only
@@ -535,13 +424,13 @@ BUILD_DEBUG(){
 echo "----------------------------------------------"
 echo " DEBUG : Debug build initiated "
 CR_TARGET=5
-CR_COMPILER=3
+CR_COMPILER=8
 CR_SELINUX=0
 CR_KSU="y"
 CR_CLEAN="n"
 echo " DEBUG : Set Build options "
-echo " DEBUG : Variant  : $CR_VARIANT_N960F"
-echo " DEBUG : Compiler : Clang 18"
+echo " DEBUG : Variant  : $CR_VARIANT_G965N"
+echo " DEBUG : Compiler : $CR_CLANG_EXP"
 echo " DEBUG : Selinux  : $CR_SELINUX Enforcing"
 echo " DEBUG : Clean    : $CR_CLEAN"
 echo "----------------------------------------------"
@@ -564,20 +453,8 @@ CR_BASE_DTB=$CR_OUTZIP/floyd/G960F-dtb
 
 # Check packages
 if ! dpkg-query -W -f='${Status}' bsdiff  | grep "ok installed"; then 
-	echo "bsdiff is missing and is required for ZIP Packaging."
-	read -p "Do you want to install bsdiff? This requires sudo privileges. (y/n) > " INSTALL_BSDIFF
-	if [ "$INSTALL_BSDIFF" = "y" ]; then
-		echo "installing bsdiff."
-		sudo apt update
-		sudo apt install -y bsdiff
-		if ! dpkg-query -W -f='${Status}' bsdiff | grep "ok installed"; then
-			echo "Failed to install bsdiff. Please try installing it manually."
-			exit 0;
-		fi
-	else
-		echo "Please install bsdiff with sudo apt install bsdiff and try again."
-		exit 0;
-	fi
+	echo " bsdiff is missing, please install with sudo apt install bsdiff" 
+	exit 0; 
 fi
 
 # Initalize with base image (Starlte)
@@ -660,19 +537,18 @@ echo "----------------------------------------------"
 read -p "Please select your build target (1-8) > " CR_TARGET
 echo "----------------------------------------------"
 echo " "
-echo "1) Clang 8 (^)"
-echo "2) Clang 12 (LLVM +LTO)"
-echo "3) Clang 14 (LLVM +LTO)"
-echo "4) Clang 18 (LLVM +LTO PGO Bolt MLGO Polly)"
-echo "5) Clang 19 (^)"
-echo "6) Neutron Clang 18 (^)"
-echo "7) Neutron Clang 19 (^)"
-echo "8) Neutron Clang 20 (BETA)"
-echo "9) Other (DS/toolchain/clang-custom)"
+echo "1) $CR_GCC4 (GCC 4.9)"
+echo "2) $CR_GCC9 (GCC 9.x)" 
+echo "3) $CR_GCC12 (GCC 12.x)" 
+echo "4) $CR_GCC13 (GCC 13.x)" 
+echo "5) $CR_CLANG_11 (Clang 11 - Samsung)"
+echo "6) $CR_CLANG_12 (Clang 12)" 
+echo "7) $CR_CLANG_14 (Clang 14)"
+echo "8) $CR_CLANG_EXP (Clang 18)"
 echo " "
-read -p "Please select your compiler (1-9) > " CR_COMPILER
+read -p "Please select your compiler (1-8) > " CR_COMPILER
 echo " "
-echo "1) SELinux Permissive "  "2) SELinux Enforcing"
+echo "1) Selinux Permissive " "2) Selinux Enforcing"
 echo " "
 read -p "Please select your SElinux mode (1-2) > " CR_SELINUX
 echo " "
@@ -685,28 +561,6 @@ fi
 echo " "
 read -p "Clean Builds? (y/n) > " CR_CLEAN
 echo " "
-
-# Validate options
-if ! [[ "$CR_TARGET" =~ ^[1-8]$ ]]; then
-    CR_TARGET=$DEFAULT_TARGET
-    echo " No target selected, defaulting to star2ltekor"
-fi
-
-if ! [[ "$CR_COMPILER" =~ ^[1-9]$ ]]; then
-    CR_COMPILER=$DEFAULT_COMPILER
-fi
-
-if ! [[ "$CR_SELINUX" =~ ^[1-2]$ ]]; then
-    CR_SELINUX=$DEFAULT_SELINUX
-fi
-
-if ! [[ "$CR_KSU" =~ ^[yYnN]$ ]]; then
-    CR_KSU=$DEFAULT_KSU
-fi
-if ! [[ "$CR_CLEAN" =~ ^[yYnN]$ ]]; then
-    CR_CLEAN=$DEFAULT_CLEAN
-fi
-
 # Call functions
 if [ "$CR_TARGET" = "7" ]; then
 echo " "
